% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosponsor_senate_15.R
\docType{data}
\name{cosponsor_senate_15}
\alias{cosponsor_senate_15}
\title{Bill cosponsorship data for the 115th Senate}
\format{
a data frame of bill cosponsorships
}
\usage{
cosponsor_senate_15
}
\description{
Bill cosponsorship data for the 115th Senate
}
\references{
govtrack.us
}
\keyword{datasets}
