% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparsity_filter.R
\name{disparsity_filter}
\alias{disparsity_filter}
\title{Disparity Filter}
\usage{
disparsity_filter(g, proj = "true", alpha = 0.05, cut_mode = "or")
}
\arguments{
\item{g}{igraph object. either two-mode or weighted network}

\item{proj}{string. Which mode to project on ("true"/"false")}

\item{alpha}{significant level}

\item{cut_mode}{'and' (retain edge if both directions are significant) or 'or' (retain edge if one direction is significant)}
}
\value{
backbone of weighted network as igraph object
}
\description{
Extract significant edges of a weighted network or one-mode projection with the disparsity filter.
}
\references{
Serrano et al. (2009). Extracting the multiscale backbone of complex weighted networks
}
\author{
David Schoch
}
