% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{graph_circ_arc}
\alias{graph_circ_arc}
\title{circular arc graph}
\usage{
graph_circ_arc(n, r, pos = 1/10, neg = 1/10, skew = 0)
}
\arguments{
\item{n}{number of nodes}

\item{r}{radius}

\item{pos}{distance fraction between positive edges}

\item{neg}{distance fraction between negative edges}

\item{skew}{create clustered graph}
}
\value{
circular arc graph as igraph object
}
\description{
circular arc graph with positive and negative ties.
}
\author{
David Schoch
}
