% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{graph_random_vote}
\alias{graph_random_vote}
\title{generate random roll-call votes based on ideology space}
\usage{
graph_random_vote(M = 101, D = 1, p = 4, pd = 2, beta = 1,
  r = 9, noprob = 0.05, Nrand = 1000, N = 525)
}
\arguments{
\item{M}{number of members}

\item{D}{distance between means}

\item{p}{dimension of space}

\item{pd}{dimensioms where distributions are separated}

\item{beta}{scaling parameter for probabilistic voting}

\item{r}{radius of hypersphere for random generation}

\item{noprob}{probabilit of non voting}

\item{Nrand}{number of randomly generated votes}

\item{N}{number of votes to sample from randomly generated votes}
}
\value{
list with random votes and ideologies
}
\description{
generate random roll-call votes based on ideology space
}
\references{
Aldrich, John H., and Montgomery, Jacob M., and Sparks, David B. (2014). Polarization and Ideology: Partisan Sources of Low Dimensionality in Scaled Roll Call Analyses. Political Analysis 22:435-456
}
\author{
David Schoch
}
