% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{graph_rectangle}
\alias{graph_rectangle}
\title{Boxicity 2 graph}
\usage{
graph_rectangle(n, r = 2, sd = 0.5)
}
\arguments{
\item{n}{number of nodes}

\item{r}{radius}

\item{sd}{standard deviation}
}
\value{
Boxicity 2 graph as igraph object
}
\description{
Create a random graph with boxicity 2.
}
\author{
David Schoch
}
