% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{graph_tolerance}
\alias{graph_tolerance}
\title{Tolerance graph}
\usage{
graph_tolerance(n, r = 2, sd = 0.5, tol = 0.5)
}
\arguments{
\item{n}{number of nodes}

\item{r}{radius}

\item{sd}{standard deviation}

\item{tol}{tolerance}
}
\value{
interval graph as igraph object and interval representation
}
\description{
Create a random interval graph.
}
\author{
David Schoch
}
