% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_filter.R
\name{hypergeometric_filter}
\alias{hypergeometric_filter}
\title{hypergeometric filter}
\usage{
hypergeometric_filter(g, type = "true", alpha = 0.05)
}
\arguments{
\item{g}{igraph object. The two-mode network}

\item{type}{which mode to project on}

\item{alpha}{significants level}
}
\value{
backbone of weighted network
}
\description{
Extract significant edges with a hypergeometric distribution.
}
\author{
David Schoch
}
