% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazarus_count.R
\name{is_interval}
\alias{is_interval}
\title{Check whether graph is interval graph}
\usage{
is_interval(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
Logical scalar, whether graph is an interval graph
}
\description{
Check whether graph is interval graph.
}
\details{
This function is not very efficient since it relies on the clique vertex matrix. More efficient linear time algorithms will be implemented in the future.
}
\author{
David Schoch
}
