% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian_vectors.R
\name{laplacian_vectors}
\alias{laplacian_vectors}
\title{laplacian Vectors}
\usage{
laplacian_vectors(g, k = 2)
}
\arguments{
\item{g}{igraph object}

\item{k}{number of vectors to return}
}
\value{
dataframe of vectors
}
\description{
Laplacian Vectors
}
\author{
David Schoch
}
