% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leveling_ties.R
\name{leveling_ties}
\alias{leveling_ties}
\title{Extract leveling ties}
\usage{
leveling_ties(g, proj = "true", thresh = 0, alpha = 2)
}
\arguments{
\item{g}{igraph object. The two-mode network}

\item{proj}{string indicating mode to project on}

\item{thresh}{weight threshold}

\item{alpha}{se addition}
}
\value{
network of leveling ties
}
\description{
Extracting leveling ties.
}
\author{
David Schoch
}
