% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexBFS.R
\name{multiLexBFS}
\alias{multiLexBFS}
\title{Multisweep Lex-BFS}
\usage{
multiLexBFS(g, k = 4)
}
\arguments{
\item{g}{igraph object}

\item{k}{number of sweeps}
}
\value{
permutation
}
\description{
Multisweep lexicograpical BFS to recoginze interval graphs.
}
\author{
David Schoch
}
