% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supergraphs.R
\name{perm2box}
\alias{perm2box}
\title{Box representation from permutations}
\usage{
perm2box(g, perm, dim)
}
\arguments{
\item{g}{igraph object.}

\item{perm}{integer vector of length n times dim}

\item{dim}{integer. dimensionality of boxes}
}
\value{
coordinates
}
\description{
Create a box representation from permutations
}
\references{
Chandran, L. S., Francis, M. C. & Sivadasan, N. Geometric representation of graphs in low dimension using axis parallel boxes. Algorithmica 56, 129.
}
