% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdsm.R
\name{sdsm}
\alias{sdsm}
\title{Stochastic Degree Sequence Model}
\usage{
sdsm(g, proj = "true", model = "logit", max_iter = 1000,
  alpha = 0.05, params = list(b0 = 0.1, b1 = 5e-05, b2 = 5e-05, b3 =
  5e-05, a = 0.01))
}
\arguments{
\item{g}{igraph object. The two-mode network}

\item{proj}{string. Which mode to project on ("true"/"false")}

\item{model}{string. which link to be used ('logit','probit','cloglog' or 'scobit')}

\item{max_iter}{number of randomly sampled networks}

\item{alpha}{significance level}

\item{params}{named parameter list for scobit model}
}
\value{
backbone of one-mode projection
}
\description{
Stochastic Degree Sequence Model.
}
\references{
Neal, Zachary (2014). The backbone of bipartite projections: Inferring relationships from co-authorship, co-sponsorship, co-attendance and other co-behaviors
}
\author{
David Schoch
}
