% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdsm.R
\name{sdsm_diagnostic}
\alias{sdsm_diagnostic}
\title{sdsm model diagnostics}
\usage{
sdsm_diagnostic(g, proj = "true", iter = 10, verbose = FALSE,
  params = list(b0 = 0.1, b1 = 5e-05, b2 = 5e-05, b3 = 5e-05, a = 0.01),
  ...)
}
\arguments{
\item{g}{igraph object. The two-mode network}

\item{proj}{string. Which mode to project on}

\item{iter}{number of fits per model}

\item{verbose}{logical. print additional information (default: FALSE)}

\item{params}{named parameter list for scobit model}

\item{...}{additional parameters passed to stats4::mle}
}
\value{
rmse and runtime of various models
}
\description{
check which link function to use
}
\author{
David Schoch
}
