% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparsity_filter.R
\name{disparsity_filter}
\alias{disparsity_filter}
\title{Disparity Filter}
\usage{
disparsity_filter(g, alpha = 0.05, cut_mode = "or")
}
\arguments{
\item{g}{igraph object. The two-mode network}

\item{alpha}{significants level}

\item{cut_mode}{'and' or 'or'}
}
\value{
backbone of weighted network
}
\description{
Extract significant edges with disparsity filter.
}
\references{
Serrano et al. (2009). Extracting the multiscale backbone of complex weighted networks
}
\author{
David Schoch
}
