% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{graph_indifference}
\alias{graph_indifference}
\title{Indifference graph}
\usage{
graph_indifference(n, r = 2)
}
\arguments{
\item{n}{number of nodes}

\item{r}{radius}
}
\value{
indifference graph as igraph object and interval representation
}
\description{
Create a random indifference graph.
}
\author{
David Schoch
}
