% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers}
\alias{helpers}
\alias{biggest_component}
\alias{clique_vertex_mat}
\alias{count_triangle_type}
\alias{is_bipartite1}
\title{helper function}
\usage{
biggest_component(g)

clique_vertex_mat(g)

count_triangle_type(g)

is_bipartite1(g)
}
\arguments{
\item{g}{igraph object.}
}
\value{
igraph object
}
\description{
small functions to deal with typical network problems
}
\author{
David Schoch
}
