% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers}
\alias{helpers}
\alias{clique_vertex_mat}
\alias{is_bipartite1}
\title{helper function}
\usage{
clique_vertex_mat(g)

is_bipartite1(g)
}
\arguments{
\item{g}{igraph object.}
}
\value{
igraph object
}
\description{
small helper functions
}
\author{
David Schoch
}
