% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian_vectors.R
\name{laplacian_vectors}
\alias{laplacian_vectors}
\title{Laplacian Vectors}
\usage{
laplacian_vectors(g, k = 2)
}
\arguments{
\item{g}{igraph object}

\item{k}{number of vectors to return}
}
\value{
data.frame of vectors
}
\description{
Returns Laplacian eigenvectors associated with the k smallest positive eigenvalues
}
\author{
David Schoch
}
