% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazarus_count.R
\name{lazarus_count}
\alias{lazarus_count}
\title{Lazarus Count}
\usage{
lazarus_count(g, perm = NULL, mode = "cols")
}
\arguments{
\item{g}{either an igraph object or a matrix}

\item{perm}{permutation or NA}

\item{mode}{one of "mcl", "cols" "rows" or "sym".}
}
\value{
Lazarus count of g
}
\description{
Calculates the Lazarus count of a matrx/graph.
}
\author{
David Schoch
}
