% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supergraphs.R
\name{superbox_graph}
\alias{superbox_graph}
\title{Supergraph with low boxicity}
\usage{
superbox_graph(g, dim)
}
\arguments{
\item{g}{igraph object}

\item{dim}{integer. target boxicity (1-5)}
}
\description{
Create supergraph with low boxicity
}
